Cleanup scripts.js by moving excess code in the event listeners to the domUpdates file
