+++ src/scripts.js
@@ -1,6 +1,6 @@
 //NOTE: Data model and non-dom manipulating logic will live in this file.
 import "./styles.css";
-import "./print-styles.css"
+import "./print-styles.css";
 import "./images/bookmark-regular.svg";
 import "./images/bookmark-solid.svg";
 import "./images/x-solid.svg";
@@ -15,9 +15,17 @@ import {
   deleteRecipe,
   displayRecipeTag,
   buildSearchFail,
+  updateActiveTags,
+  updateUser,
+  updateActiveRecipes,
 } from "./domUpdates.js";
 
-import { filterByTag, searchRecipes, getIngredientNames, calculateCost } from "../src/recipes.js";
+import {
+  filterByTag,
+  searchRecipes,
+  getIngredientNames,
+  calculateCost,
+} from "../src/recipes.js";
 const activeTags = [];
 let currentUser;
 let activeRecipes;
@@ -38,31 +46,24 @@ const discoverRecipes = document.querySelector("#discoverRecipes");
 // ===== EVENT LISTENERS =====
 
 window.addEventListener("load", function () {
-  Promise.all(promises).then((res) => {
-    data = {
-      users: res["0"].users,
-      ingredients: res["1"].ingredients,
-      recipes: res["2"].recipes,
-    };
+  Promise.all(promises)
+    .then((res) => {
+      data = {
+        users: res["0"].users,
+        ingredients: res["1"].ingredients,
+        recipes: res["2"].recipes,
+      };
 
-    loadUser(res["0"].users);
-   
-    createRecipeCards(res["2"].recipes);
-    activeRecipes = [...res["2"].recipes];
-  });
+      loadUser(res["0"].users);
+      createRecipeCards(res["2"].recipes);
+      activeRecipes = [...res["2"].recipes];
+    })
+    .catch((err) => console.log(err));
 });
 
 tagSection.addEventListener("click", function (event) {
   let tag = event.target.closest(".tag-card");
-  tag.classList.toggle("tag-active");
-
-  let tagId = tag.id;
-  if (!activeTags.includes(tagId)) {
-    activeTags.push(tagId);
-  } else {
-    let index = activeTags.indexOf(tagId);
-    activeTags.splice(index, 1);
-  }
+  updateActiveTags(tag, activeTags);
   let filteredArray = filterByTag(activeTags, activeRecipes);
   createRecipeCards(filteredArray);
 });
@@ -70,7 +71,6 @@ tagSection.addEventListener("click", function (event) {
 searchButton.addEventListener("click", function (event) {
   let searchTerm = searchInput.value;
   let searchedRecipes = searchRecipes(searchTerm, activeRecipes);
-
   createRecipeCards(searchedRecipes);
   if (searchedRecipes.length === 0) {
     buildSearchFail();
@@ -81,7 +81,7 @@ recipeArea.addEventListener("click", function (event) {
   let recipeClicked = event.target.parentElement.id;
   let foundRecipe = locateRecipe(recipeClicked, data.recipes);
   let recipeIngredients = getIngredientNames(foundRecipe, data.ingredients);
-  let cost = calculateCost(foundRecipe, data.ingredients)
+  let cost = calculateCost(foundRecipe, data.ingredients);
   buildRecipeCard(foundRecipe, data.ingredients, recipeIngredients, cost);
   displayRecipeTag(recipeClicked, currentUser, data.recipes);
   displayRecipeCard();
@@ -95,21 +95,10 @@ recipeCardClose.addEventListener("click", function (event) {
 recipeCardBookmarkAdd.addEventListener("click", function (event) {
   let bookmarkClicked = event.target.id;
   addRecipe(currentUser.id, bookmarkClicked).then((responseData) => {
-    console.log("Data from the POST request: ", responseData);
     data.users = responseData;
     let users = data.users.users;
-    const updateUser = (users) => {
-      return users.find((user) => {
-        if (user.id === currentUser.id) {
-          return user;
-        }
-      });
-    };
-    currentUser = updateUser(users);
-    // updateUser updates the currentUser with the data from the database!!!
-    console.log(currentUser);
-    // currentUser.id => the place where data.users (id matches)
-    // saveRecipe(bookmarkClicked, currentUser);
+    currentUser = updateUser(users, currentUser);
+    // updateUser updates the currentUser with the data from the database not the local DM!!!
     displayRecipeTag(bookmarkClicked, currentUser, data.recipes);
   });
 });
@@ -118,19 +107,11 @@ recipeCardBookmarkDelete.addEventListener("click", function (event) {
   let bookmarkClicked = event.target.id;
   deleteRecipe(bookmarkClicked, currentUser);
   displayRecipeTag(bookmarkClicked, currentUser, data.recipes);
-  let recipesToCook = currentUser.recipesToCook.map((recipeId) => {
-    let wholeRecipe = locateRecipe(recipeId, data.recipes);
-    return wholeRecipe;
-  });
-  activeRecipes = recipesToCook;
+  activeRecipes = updateActiveRecipes(currentUser, data);
 });
 
 userSavedRecipes.addEventListener("click", function (event) {
-  let recipesToCook = currentUser.recipesToCook.map((recipeId) => {
-    let wholeRecipe = locateRecipe(recipeId, data.recipes);
-    return wholeRecipe;
-  });
-  activeRecipes = recipesToCook;
+  activeRecipes = updateActiveRecipes(currentUser, data);
   createRecipeCards(activeRecipes);
 });
